asset-pipeline/requirements.txt
gradio>=4.44.0
Pillow>=10.0.0
trimesh>=4.0.0
numpy>=1.24.0
huggingface-hub>=0.20.0
